+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # AvalynxLoader Changelog
 
+## 0.0.2
+- [x] `AvalynxLoader` constructor now accepts a `language` parameter
+- [x] `loaderText` moved from options to language
+
 ## 0.0.1
 - [x] Initial release